backend/requirements.txt
configargparse
redis
paho-mqtt
